OIE_toolkit/requirements.txt
pyyaml>=6.0
jupyterlab>=4.0
transformers>=4.43.0
torch>=2.2.0
